mv all models to model pkg mv all consts to consts.go
