+++ model/models.go
@@ -1,9 +1,12 @@
 package model
 
 import (
+	"github.com/koupleless/virtual-kubelet/tunnel"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/client-go/kubernetes"
+	"sigs.k8s.io/controller-runtime/pkg/cache"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"time"
 )
 
@@ -12,17 +15,6 @@ type NetworkInfo struct {
 	NodeIP   string `json:"nodeIP"`
 	HostName string `json:"hostName"`
 }
-
-// NodeStatus is the node curr status
-type NodeStatus string
-
-const (
-	// NodeStatusActivated node activated, will start vnode if not being started
-	NodeStatusActivated NodeStatus = "ACTIVATED"
-
-	// NodeStatusDeactivated node deactivated, will shut down vnode if started
-	NodeStatusDeactivated NodeStatus = "DEACTIVATED"
-)
 
 // NodeMetadata is the base data of a vnode, will be transfer to default labels of a vnode
 type NodeMetadata struct {
@@ -63,26 +55,6 @@ type NodeStatusData struct {
 	CustomConditions  []v1.NodeCondition               `json:"customConditions"`
 }
 
-// OperationResponseResult is the container operation response result
-type OperationResponseResult string
-
-const (
-	OperationResponseCodeSuccess OperationResponseResult = "SUCCESS"
-	OperationResponseCodeFailure OperationResponseResult = "FAIL"
-)
-
-// ContainerState is the state of a container, will set to pod state and show on k8s
-type ContainerState string
-
-const (
-	ContainerStateActivated   = "ACTIVATED"
-	ContainerStateResolved    = "RESOLVED"
-	ContainerStateDeactivated = "DEACTIVATED"
-)
-
-// PodKeyAll present container status will share to all pods
-const PodKeyAll = "all"
-
 // ContainerStatusData is the status data of a container
 type ContainerStatusData struct {
 	// Key generated by tunnel, need to be the same as Tunnel GetContainerUniqueKey of same container
@@ -103,4 +75,68 @@ type K8SConfig struct {
 
 	// InformerSyncPeriod
 	InformerSyncPeriod time.Duration
+}
+
+type BuildVNodeProviderConfig struct {
+	// NodeIP is the ip of the node
+	NodeIP string `json:"nodeIP"`
+
+	// NodeHostname is the hostname of the node
+	NodeHostname string `json:"nodeHostname"`
+
+	// Name is the node name, will be sent to utils.FormatNodeName to construct vnode name, and Name will be set to label
+	Name string `json:"name"`
+
+	// Name is the node version, will be set to label
+	Version string `json:"version"`
+
+	// ENV is the env of node, will be set to label
+	Env string `json:"env"`
+
+	// CustomTaints is the taints set by tunnel
+	CustomTaints []v1.Taint `json:"customTaints"`
+}
+
+type BuildVNodeConfig struct {
+	// Client is the runtime client instance
+	Client client.Client
+
+	// KubeCache is the cache of kube resources
+	KubeCache cache.Cache
+
+	// Tunnel is the tunnel of container and node management
+	Tunnel tunnel.Tunnel
+
+	// NodeID is the unique id of node, should be unique key of system
+	NodeID string
+
+	// NodeIP is the ip of node
+	NodeIP string
+
+	// NodeHostname is the hostname of node
+	NodeHostname string
+
+	// NodeName is the name of node, will set to node's label
+	NodeName string
+
+	// NodeVersion is the version of node, will set to node's label
+	NodeVersion string
+
+	// Env is the runtime env of virtual-kubelet, will set to node created by virtual kubelet
+	Env string
+
+	// CustomTaints is the taint set by tunnel
+	CustomTaints []v1.Taint
+}
+
+type BuildVNodeControllerConfig struct {
+	ClientID string
+
+	Env string
+
+	// VPodIdentity is the vpod special value of model.LabelKeyOfComponent
+	VPodIdentity string
+
+	// Tunnels is the tunnel provider supported
+	Tunnels []tunnel.Tunnel
 }
